Move TestTimeouts to the correct test file.
