+++ api/handlers_e2e_test.go
@@ -9,7 +9,6 @@ import (
 	"time"
 
 	"github.com/labstack/echo/v4"
-	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
 
 	"github.com/algorand/go-algorand-sdk/encoding/json"
@@ -19,7 +18,6 @@ import (
 
 	"github.com/algorand/indexer/api/generated/v2"
 	"github.com/algorand/indexer/idb"
-	"github.com/algorand/indexer/idb/mocks"
 	"github.com/algorand/indexer/idb/postgres"
 	pgtest "github.com/algorand/indexer/idb/postgres/testing"
 	"github.com/algorand/indexer/util/test"
@@ -334,184 +332,4 @@ func TestVersion(t *testing.T) {
 	require.False(t, response.IsMigrating)
 	// This is weird looking because the version is set with -ldflags
 	require.Equal(t, response.Version, "(unknown version)")
-}
-
-func TestTimeouts(t *testing.T) {
-	// function pointers to execute the different DB operations. We really only
-	// care that they timeout with WaitUntil, but the return arguments need to
-	// be correct to avoid a panic.
-	mostMockFunctions := func(method string) func(mockIndexer *mocks.IndexerDb, timeout <-chan time.Time) {
-		return func(mockIndexer *mocks.IndexerDb, timeout <-chan time.Time) {
-			mockIndexer.
-				On(method, mock.Anything, mock.Anything, mock.Anything).
-				WaitUntil(timeout).
-				Return(nil, uint64(0))
-		}
-	}
-	transactionFunc := mostMockFunctions("Transactions")
-	applicationsFunc := mostMockFunctions("Applications")
-	accountsFunc := mostMockFunctions("GetAccounts")
-	assetsFunc := mostMockFunctions("Assets")
-	balancesFunc := mostMockFunctions("AssetBalances")
-	blockFunc := func(mockIndexer *mocks.IndexerDb, timeout <-chan time.Time) {
-		mockIndexer.
-			On("GetBlock", mock.Anything, mock.Anything, mock.Anything).
-			WaitUntil(timeout).
-			Return(bookkeeping.BlockHeader{}, nil, nil)
-	}
-	healthFunc := func(mockIndexer *mocks.IndexerDb, timeout <-chan time.Time) {
-		mockIndexer.
-			On("Health", mock.Anything, mock.Anything, mock.Anything).
-			WaitUntil(timeout).
-			Return(idb.Health{}, nil)
-	}
-
-	// Call each of the handlers and let the database timeout.
-	testcases := []struct {
-		name        string
-		errString   string
-		mockCall    func(mockIndexer *mocks.IndexerDb, timeout <-chan time.Time)
-		callHandler func(ctx echo.Context, si ServerImplementation) error
-	}{
-		{
-			name:      "SearchForTransactions",
-			errString: errTransactionSearch,
-			mockCall:  transactionFunc,
-			callHandler: func(ctx echo.Context, si ServerImplementation) error {
-				return si.SearchForTransactions(ctx, generated.SearchForTransactionsParams{})
-			},
-		},
-		{
-			name:      "LookupAccountTransactions",
-			errString: errTransactionSearch,
-			mockCall:  transactionFunc,
-			callHandler: func(ctx echo.Context, si ServerImplementation) error {
-				return si.LookupAccountTransactions(ctx, "", generated.LookupAccountTransactionsParams{})
-			},
-		},
-		{
-			name:      "LookupAssetTransactions",
-			errString: errTransactionSearch,
-			mockCall:  transactionFunc,
-			callHandler: func(ctx echo.Context, si ServerImplementation) error {
-				return si.LookupAssetTransactions(ctx, 1, generated.LookupAssetTransactionsParams{})
-			},
-		},
-		{
-			name:      "LookupApplicaitonLogsByID",
-			errString: errTransactionSearch,
-			mockCall:  transactionFunc,
-			callHandler: func(ctx echo.Context, si ServerImplementation) error {
-				return si.LookupApplicationLogsByID(ctx, 1, generated.LookupApplicationLogsByIDParams{})
-			},
-		},
-		{
-			name:      "LookupApplicationByID",
-			errString: errFailedSearchingApplication,
-			mockCall:  applicationsFunc,
-			callHandler: func(ctx echo.Context, si ServerImplementation) error {
-				return si.LookupApplicationByID(ctx, 0, generated.LookupApplicationByIDParams{})
-			},
-		},
-		{
-			name:      "SearchForApplications",
-			errString: errFailedSearchingApplication,
-			mockCall:  applicationsFunc,
-			callHandler: func(ctx echo.Context, si ServerImplementation) error {
-				return si.SearchForApplications(ctx, generated.SearchForApplicationsParams{})
-			},
-		},
-		{
-			name:      "SearchForAccount",
-			errString: errFailedSearchingAccount,
-			mockCall:  accountsFunc,
-			callHandler: func(ctx echo.Context, si ServerImplementation) error {
-				return si.SearchForAccounts(ctx, generated.SearchForAccountsParams{})
-			},
-		},
-		{
-			name:      "LookupAccountByID",
-			errString: errFailedSearchingAccount,
-			mockCall:  accountsFunc,
-			callHandler: func(ctx echo.Context, si ServerImplementation) error {
-				return si.LookupAccountByID(ctx,
-					"PBH2JQNVP5SBXLTOWNHHPGU6FUMBVS4ZDITPK5RA5FG2YIIFS6UYEMFM2Y",
-					generated.LookupAccountByIDParams{})
-			},
-		},
-		{
-			name:      "SearchForAssets",
-			errString: errFailedSearchingAsset,
-			mockCall:  assetsFunc,
-			callHandler: func(ctx echo.Context, si ServerImplementation) error {
-				return si.SearchForAssets(ctx, generated.SearchForAssetsParams{})
-			},
-		},
-		{
-			name:      "LookupAssetByID",
-			errString: errFailedSearchingAsset,
-			mockCall:  assetsFunc,
-			callHandler: func(ctx echo.Context, si ServerImplementation) error {
-				return si.LookupAssetByID(ctx, 1, generated.LookupAssetByIDParams{})
-			},
-		},
-		{
-			name:      "LookupAssetBalances",
-			errString: errFailedSearchingAssetBalances,
-			mockCall:  balancesFunc,
-			callHandler: func(ctx echo.Context, si ServerImplementation) error {
-				return si.LookupAssetBalances(ctx, 1, generated.LookupAssetBalancesParams{})
-			},
-		},
-		{
-			name:      "LookupBlock",
-			errString: errLookingUpBlock,
-			mockCall:  blockFunc,
-			callHandler: func(ctx echo.Context, si ServerImplementation) error {
-				return si.LookupBlock(ctx, 100)
-			},
-		},
-		{
-			name:      "Health",
-			errString: errFailedLookingUpHealth,
-			mockCall:  healthFunc,
-			callHandler: func(ctx echo.Context, si ServerImplementation) error {
-				return si.MakeHealthCheck(ctx)
-			},
-		},
-	}
-
-	for _, tc := range testcases {
-		t.Run(tc.name, func(t *testing.T) {
-			timeout := make(chan time.Time, 1)
-			defer func() {
-				timeout <- time.Now()
-				close(timeout)
-			}()
-
-			// Make a mock indexer and tell the mock to timeout.
-			mockIndexer := &mocks.IndexerDb{}
-
-			si := ServerImplementation{
-				db:      mockIndexer,
-				timeout: 5 * time.Millisecond,
-			}
-
-			// Setup context...
-			e := echo.New()
-			req := httptest.NewRequest(http.MethodGet, "/", nil)
-			rec1 := httptest.NewRecorder()
-			c := e.NewContext(req, rec1)
-
-			// configure the mock to timeout, then call the handler.
-			tc.mockCall(mockIndexer, timeout)
-			err := tc.callHandler(c, si)
-
-			require.NoError(t, err)
-			bodyStr := rec1.Body.String()
-			require.Equal(t, 500, rec1.Code)
-			require.Contains(t, bodyStr, tc.errString)
-			require.Contains(t, bodyStr, "timeout")
-		})
-	}
 }
